Add "tagline" and "hidden-on-frontpage" attributes to activities.
Taglines are now displayed in font similar to h2 and just below the
h1 activity name.

Activities that are "hidden-on-frontpage" are not included in the
activity listing on the frontpage (but are still included on the
/activities/ page for transparency's sake). Intended use is to not
confuse visitors by listing both eg tryggve and tryggve2, but
instead guide visitors to user-facing landing pages like /tryggve/
(which then links to /tryggve1/ and /tryggve2/ which holds
administrivia for those particular projects).