Enabled site-wide RSS feed.
Using jekyll-feed plugin which is recommended by Github Pages
https://help.github.com/en/articles/atom-rss-feeds-for-github-pages

Did not use the {% feed_meta %} tag in head.html, instead created the RSS <link> manually
because feed_meta always messed up the path to the xml file for some reason (have not confirmed
this, but likely due to other settings in _config.yml). Anyway, this way should be just as robust.